requirements-dev.txt
# Packaging + Distribution
wheel

# Base Dependencies
dynaconf
grpcio
grpcio-tools
grpcio-reflection


# Testing
ipython
ipdb
pytest
pytest-cov

# Linting + Formatting
ruff